feat: move wallet provider to wallet-sdk
