+++ packages/dapp-kit/src/hooks/index.ts
@@ -1,6 +1,5 @@
 export * from './useHubProvider';
 export * from './useSpokeProvider';
-export * from './useWalletProvider';
 export * from './useSuppliedAssets';
 export * from './useHubWalletAddress';
 export * from './useSupply';
