+++ packages/sdk/src/entities/icon/IconSpokeProvider.ts
@@ -1,6 +1,6 @@
 import type { IconSpokeChainConfig } from '../../types.js';
-import type { IIconWalletProvider } from '../../interfaces.js';
 import IconService from 'icon-sdk-js';
+import type { IIconWalletProvider } from '@sodax/types';
 
 export class IconSpokeProvider {
   public readonly walletProvider: IIconWalletProvider;
