+++ packages/sdk/src/services/spoke/SolanaSpokeService.ts
@@ -9,7 +9,8 @@ import { getAssetManagerProgram, getConnectionProgram } from '../../entities/sol
 import type { SolanaSpokeProvider } from '../../entities/solana/SolanaSpokeProvider.js';
 import { AssetManagerPDA, ConnectionConfigPDA } from '../../entities/solana/pda/pda.js';
 import { isNative } from '../../entities/solana/utils/utils.js';
-import type { HubAddress, PromiseSolanaTxReturnType, SolanaReturnType } from '../../types.js';
+import type { PromiseSolanaTxReturnType, SolanaReturnType } from '../../types.js';
+import type { HubAddress } from '@sodax/types';
 import { EvmWalletAbstraction } from '../hub/index.js';
 
 export type SolanaSpokeDepositParams = {
@@ -28,7 +29,7 @@ export type TransferToHubParams = {
 };
 
 export class SolanaSpokeService {
-  private constructor() {}
+  private constructor() { }
 
   public static async deposit<R extends boolean = false>(
     params: SolanaSpokeDepositParams,
