+++ packages/wallet-sdk/package.json
@@ -6,7 +6,10 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "module": "dist/index.mjs",
-  "files": ["dist", "src"],
+  "files": [
+    "dist",
+    "src"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/icon-project/sodax-frontend"
@@ -23,7 +26,6 @@
     "lint": "biome lint . --write"
   },
   "dependencies": {
-    "@sodax/types": "workspace:*",
     "@archwayhq/arch3.js": "0.7.0",
     "@coral-xyz/anchor": "0.30.1",
     "@cosmjs/cosmwasm-stargate": "0.32.2",
@@ -38,6 +40,8 @@
     "@keplr-wallet/types": "0.12.18",
     "@mysten/dapp-kit": "0.14.18",
     "@mysten/sui": "^1.21.2",
+    "@mysten/wallet-standard": "^0.15.6",
+    "@sodax/types": "workspace:*",
     "@solana/wallet-adapter-react": "0.15.35",
     "@solana/wallet-adapter-wallets": "0.19.32",
     "@solana/web3.js": "1.98.0",
