+++ packages/wallet-sdk/src/hooks/index.ts
@@ -11,3 +11,4 @@ export { useXService } from './useXService';
 export { useXBalances } from './useXBalances';
 export { useWalletProviderOptions } from './useWalletProviderOptions';
 export { useEvmSwitchChain } from './useEvmSwitchChain';
+export { useWalletProvider } from './useWalletProvider';
