+++ packages/wallet-sdk/src/index.ts
@@ -15,6 +15,7 @@ export * from './xchains/sui';
 export * from './hooks';
 export * from './useXWagmiStore';
 export * from './XWagmiProviders';
+export * from './wallet-providers';
 
 export * from './types';
 export type * from './types';
